Adding support for macos frameworks and libaries without dSYM
If a library is from within a frammework, the dSYM is not in the
same folder than the library, but has an own special
.framework.dSYM folder
Try to resolve this folder if no dSYM was found next to the executable.

If no dSYM was found, try to load the information attached to the
library. This make sure, we atleast get the symbol table and
can try to resolve some function names.

Using the vmslide from _dyld_get_image_vmaddr_slide doesn't work
for frameworks. Instead of calculating the image_base_address ourself
use _dyld_get_image_header to get the already calculated base address,
which also works for frameworks.